+++ openTEPES/openTEPES_ProblemSolving.py
@@ -1,5 +1,5 @@
 """
-Open Generation, Storage, and Transmission Operation and Expansion Planning Model with RES and ESS (openTEPES) - October 26, 2023
+Open Generation, Storage, and Transmission Operation and Expansion Planning Model with RES and ESS (openTEPES) - December 09, 2023
 """
 
 import time
@@ -131,18 +131,18 @@ def ProblemSolving(DirName, CaseName, SolverName, OptModel, mTEPES, pIndLogConso
     print    ('  Problem size                         ... ', OptModel.model().nconstraints(), 'constraints, ', OptModel.model().nvariables()-mTEPES.nFixedVariables+1, 'variables')
     print    ('  Solution time                        ... ', round(SolvingTime), 's')
     print    ('  Total system                 cost [MEUR] ', OptModel.vTotalSCost())
-    print    ('  Total generation  investment cost [MEUR] ', sum(mTEPES.pDiscountFactor[p] * mTEPES.pGenInvestCost [gc      ]   * OptModel.vGenerationInvest[p,gc      ]() for gc       in mTEPES.gc))
-    print    ('  Total generation  retirement cost [MEUR] ', sum(mTEPES.pDiscountFactor[p] * mTEPES.pGenRetireCost [gd      ]   * OptModel.vGenerationRetire[p,gd      ]() for gd       in mTEPES.gd))
+    print    ('  Total generation  investment cost [MEUR] ', sum(mTEPES.pDiscountedWeight[p] * mTEPES.pGenInvestCost [gc      ]   * OptModel.vGenerationInvest[p,gc      ]() for gc       in mTEPES.gc))
+    print    ('  Total generation  retirement cost [MEUR] ', sum(mTEPES.pDiscountedWeight[p] * mTEPES.pGenRetireCost [gd      ]   * OptModel.vGenerationRetire[p,gd      ]() for gd       in mTEPES.gd))
     if mTEPES.pIndHydroTopology == 1 and len(mTEPES.rn):
-        print('  Total reservoir   investment cost [MEUR] ', sum(mTEPES.pDiscountFactor[p] * mTEPES.pRsrInvestCost [rc      ]   * OptModel.vReservoirInvest [p,rc      ]() for rc       in mTEPES.rn))
+        print('  Total reservoir   investment cost [MEUR] ', sum(mTEPES.pDiscountedWeight[p] * mTEPES.pRsrInvestCost [rc      ]   * OptModel.vReservoirInvest [p,rc      ]() for rc       in mTEPES.rn))
     else:
         print('  Total reservoir   investment cost [MEUR] ', 0.0)
-    print    ('  Total network     investment cost [MEUR] ', sum(mTEPES.pDiscountFactor[p] * mTEPES.pNetFixedCost  [ni,nf,cc]   * OptModel.vNetworkInvest   [p,ni,nf,cc]() for ni,nf,cc in mTEPES.lc))
+    print    ('  Total network     investment cost [MEUR] ', sum(mTEPES.pDiscountedWeight[p] * mTEPES.pNetFixedCost  [ni,nf,cc]   * OptModel.vNetworkInvest   [p,ni,nf,cc]() for ni,nf,cc in mTEPES.lc))
     if mTEPES.pIndHydrogen      == 1 and len(mTEPES.pc):
-        print('  Total pipeline    investment cost [MEUR] ', sum(mTEPES.pDiscountFactor[p] * mTEPES.pLineInvestCost[ni,nf,cc]   * OptModel.vPipelineInvest  [p,ni,nf,cc]() for ni,nf,cc in mTEPES.pc))
+        print('  Total pipeline    investment cost [MEUR] ', sum(mTEPES.pDiscountedWeight[p] * mTEPES.pLineInvestCost[ni,nf,cc]   * OptModel.vPipelineInvest  [p,ni,nf,cc]() for ni,nf,cc in mTEPES.pc))
     else:
         print('  Total pipeline    investment cost [MEUR] ', 0.0)
-    print    ('  Total generation  operation  cost [MEUR] ', sum(mTEPES.pDiscountFactor[p] * mTEPES.pScenProb      [p,sc    ]() * OptModel.vTotalGCost      [p,sc,n    ]() for n        in mTEPES.n ))
-    print    ('  Total consumption operation  cost [MEUR] ', sum(mTEPES.pDiscountFactor[p] * mTEPES.pScenProb      [p,sc    ]() * OptModel.vTotalCCost      [p,sc,n    ]() for n        in mTEPES.n ))
-    print    ('  Total emission               cost [MEUR] ', sum(mTEPES.pDiscountFactor[p] * mTEPES.pScenProb      [p,sc    ]() * OptModel.vTotalECost      [p,sc,n    ]() for n        in mTEPES.n ))
-    print    ('  Total reliability            cost [MEUR] ', sum(mTEPES.pDiscountFactor[p] * mTEPES.pScenProb      [p,sc    ]() * OptModel.vTotalRCost      [p,sc,n    ]() for n        in mTEPES.n ))
+    print    ('  Total generation  operation  cost [MEUR] ', sum(mTEPES.pDiscountedWeight[p] * mTEPES.pScenProb      [p,sc    ]() * OptModel.vTotalGCost      [p,sc,n    ]() for n        in mTEPES.n ))
+    print    ('  Total consumption operation  cost [MEUR] ', sum(mTEPES.pDiscountedWeight[p] * mTEPES.pScenProb      [p,sc    ]() * OptModel.vTotalCCost      [p,sc,n    ]() for n        in mTEPES.n ))
+    print    ('  Total emission               cost [MEUR] ', sum(mTEPES.pDiscountedWeight[p] * mTEPES.pScenProb      [p,sc    ]() * OptModel.vTotalECost      [p,sc,n    ]() for n        in mTEPES.n ))
+    print    ('  Total reliability            cost [MEUR] ', sum(mTEPES.pDiscountedWeight[p] * mTEPES.pScenProb      [p,sc    ]() * OptModel.vTotalRCost      [p,sc,n    ]() for n        in mTEPES.n ))